+++ src/VsDebugMcp.Protocol/Class1.cs
@@ -61,6 +61,8 @@ public static class BridgeMethods
 	public const string GetOutputPanes = "getOutputPanes";
 	public const string DebuggerFindSolutionProcesses = "debuggerFindSolutionProcesses";
 	public const string DebuggerAutoAttach = "debuggerAutoAttach";
+	public const string DebuggerGetSnapshot = "debuggerGetSnapshot";
+	public const string DebuggerReadMemory = "debuggerReadMemory";
 	public const string Shutdown = "shutdown";
 }
 
@@ -112,6 +114,8 @@ public static class BridgeErrorCodes
 	public const string CannotSwitchConfigurationWhileDebugging = "cannot_switch_configuration_while_debugging";
 	public const string EngineNotFound = "engine_not_found";
 	public const string NoSolutionProcessesFound = "no_solution_processes_found";
+	public const string InvalidMemoryAddress = "invalid_memory_address";
+	public const string MemoryReadFailed = "memory_read_failed";
 	public const string InternalError = "internal_error";
 }